+++ src/Middleware/TypedProperties.php
@@ -9,12 +9,31 @@ use JsonMapper\Enums\Visibility;
 use JsonMapper\JsonMapperInterface;
 use JsonMapper\ValueObjects\PropertyMap;
 use JsonMapper\Wrapper\ObjectWrapper;
+use Psr\SimpleCache\CacheInterface;
 
 class TypedProperties extends AbstractMiddleware
 {
-    public function handle(\stdClass $json, ObjectWrapper $object, PropertyMap $map, JsonMapperInterface $mapper): void
+    /** @var CacheInterface */
+    private $cache;
+
+    public function __construct(CacheInterface $cache)
     {
+        $this->cache = $cache;
+    }
+
+    public function handle(\stdClass $json, ObjectWrapper $object, PropertyMap $propertyMap, JsonMapperInterface $mapper): void
+    {
+        $propertyMap->merge($this->fetchPropertyMapForObject($object));
+    }
+
+    private function fetchPropertyMapForObject(ObjectWrapper $object): PropertyMap
+    {
+        if ($this->cache->has($object->getName())) {
+            return $this->cache->get($object->getName());
+        }
+
         $reflectionProperties = $object->getReflectedObject()->getProperties();
+        $intermediatePropertyMap = new PropertyMap();
 
         foreach ($reflectionProperties as $reflectionProperty) {
             $type = $reflectionProperty->getType();
@@ -23,13 +42,17 @@ class TypedProperties extends AbstractMiddleware
                 continue;
             }
 
-            $reflectionProperty = PropertyBuilder::new()
+            $property = PropertyBuilder::new()
                 ->setName($reflectionProperty->getName())
                 ->setType($type->getName())
                 ->setIsNullable($type->allowsNull())
                 ->setVisibility(Visibility::fromReflectionProperty($reflectionProperty))
                 ->build();
-            $map->addProperty($reflectionProperty);
+            $intermediatePropertyMap->addProperty($property);
         }
+
+        $this->cache->set($object->getName(), $intermediatePropertyMap);
+
+        return $intermediatePropertyMap;
     }
 }