+++ tests/Unit/JsonMapperTest.php
@@ -48,7 +48,7 @@ class JsonMapperTest extends TestCase
             public function handle(
                 \stdClass $json,
                 ObjectWrapper $object,
-                PropertyMap $map,
+                PropertyMap $propertyMap,
                 JsonMapperInterface $mapper
             ): void {
                 $this->called = true;